requirements.txt
tensorflow-macos==2.15.0
numpy==1.24.3
matplotlib==3.8.2
pandas==2.1.4
scikit-learn==1.3.2 